# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CampusEats API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  vendor: "vendor",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  phone?: string;
  role: UserRole;
  createdAt?: string;
}

export interface Vendor {
  id: number;
  userId?: number;
  stallName: string;
  description?: string;
  mpesaNumber?: string;
  isActive: boolean;
  imageUrl?: string;
  location?: string;
  pickupTimeMin?: number;
  pickupTimeMax?: number;
  createdAt?: string;
}

export interface MenuItem {
  id: number;
  vendorId: number;
  name: string;
  description?: string;
  price: number;
  category?: string;
  imageUrl?: string;
  isAvailable: boolean;
  orderCount?: number;
  createdAt?: string;
}

export type MenuItemWithVendor = MenuItem & {
  vendor?: Vendor;
};

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  paid: "paid",
  preparing: "preparing",
  ready: "ready",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  studentId: number;
  vendorId: number;
  studentName?: string;
  status: OrderStatus;
  totalAmount: number;
  pickupCode?: string;
  notes?: string;
  createdAt?: string;
  updatedAt?: string;
}

export interface OrderItem {
  id: number;
  orderId: number;
  menuItemId: number;
  menuItemName: string;
  quantity: number;
  unitPrice: number;
}

export type OrderWithItems = Order & {
  items?: OrderItem[];
  vendor?: Vendor;
};

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface Transaction {
  id: number;
  orderId: number;
  amount: number;
  commission?: number;
  vendorPayout?: number;
  status: TransactionStatus;
  createdAt?: string;
}

export interface MarketplaceStats {
  totalItems: number;
  totalVendors: number;
  ordersToday: number;
  avgPickupTime: number;
}

export interface VendorStats {
  totalOrders: number;
  ordersToday: number;
  totalEarnings: number;
  earningsToday: number;
  pendingOrders: number;
  completedOrders: number;
}

export interface AdminStats {
  totalOrders: number;
  totalVendors: number;
  totalRevenue: number;
  totalCommission: number;
  ordersToday: number;
  activeVendors: number;
}

export interface CreateMenuItemBody {
  vendorId: number;
  name: string;
  description?: string;
  price: number;
  category: string;
  imageUrl?: string;
  isAvailable?: boolean;
}

export interface UpdateMenuItemBody {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  imageUrl?: string;
  isAvailable?: boolean;
}

export interface CreateVendorBody {
  userId?: number;
  stallName: string;
  description?: string;
  mpesaNumber: string;
  imageUrl?: string;
  location?: string;
  pickupTimeMin?: number;
  pickupTimeMax?: number;
}

export interface UpdateVendorBody {
  stallName?: string;
  description?: string;
  mpesaNumber?: string;
  imageUrl?: string;
  location?: string;
  isActive?: boolean;
  pickupTimeMin?: number;
  pickupTimeMax?: number;
}

export type CreateOrderBodyItemsItem = {
  menuItemId: number;
  quantity: number;
};

export interface CreateOrderBody {
  studentId?: number;
  vendorId: number;
  items: CreateOrderBodyItemsItem[];
  notes?: string;
  studentName?: string;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  paid: "paid",
  preparing: "preparing",
  ready: "ready",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  student: "student",
  vendor: "vendor",
  admin: "admin",
} as const;

export interface CreateUserBody {
  name: string;
  email: string;
  phone?: string;
  password: string;
  role?: CreateUserBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface LoginResponse {
  user: User;
  token: string;
}

export type GetMarketplaceFeedParams = {
  category?: string;
  search?: string;
  limit?: number;
  offset?: number;
};

export type GetMarketplaceFeed200 = {
  items: MenuItemWithVendor[];
  total: number;
};

export type ListMenuItemsParams = {
  vendor_id?: number;
  category?: string;
  is_available?: boolean;
};

export type GetVendorOrdersParams = {
  status?: GetVendorOrdersStatus;
};

export type GetVendorOrdersStatus =
  (typeof GetVendorOrdersStatus)[keyof typeof GetVendorOrdersStatus];

export const GetVendorOrdersStatus = {
  pending: "pending",
  paid: "paid",
  preparing: "preparing",
  ready: "ready",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ListOrdersParams = {
  student_id?: number;
  status?: string;
};

export type AdminListOrdersParams = {
  status?: string;
  limit?: number;
};