# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CampusEats API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /marketplace/feed:
    get:
      operationId: getMarketplaceFeed
      tags: [marketplace]
      summary: Get unified marketplace food feed
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 40
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Marketplace food feed
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/MenuItemWithVendor"
                  total:
                    type: integer
                required: [items, total]
...
  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all available menu items
      parameters:
        - in: query
          name: vendor_id
          schema:
            type: integer
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: is_available
          schema:
            type: boolean
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a new menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Created menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a menu item by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
    put:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemBody"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: List all orders (admin view)
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithItems"
  /admin/vendors/{id}/toggle:
    put:
      operationId: toggleVendorStatus
      tags: [admin]
      summary: Toggle vendor active status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
components:
  schemas:
...
    Vendor:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        stallName:
          type: string
        description:
          type: string
        mpesaNumber:
          type: string
        isActive:
          type: boolean
        imageUrl:
          type: string
        location:
          type: string
        pickupTimeMin:
          type: integer
        pickupTimeMax:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, stallName, isActive]
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        isAvailable:
          type: boolean
        orderCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, vendorId, name, price, isAvailable]
    MenuItemWithVendor:
      allOf:
        - $ref: "#/components/schemas/MenuItem"
        - type: object
          properties:
            vendor:
              $ref: "#/components/schemas/Vendor"
    Order:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        vendorId:
          type: integer
        studentName:
          type: string
        status:
          type: string
          enum: [pending, paid, preparing, ready, completed, cancelled]
        totalAmount:
          type: number
        pickupCode:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, studentId, vendorId, status, totalAmount]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
      required: [id, orderId, menuItemId, menuItemName, quantity, unitPrice]
    OrderWithItems:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
            vendor:
              $ref: "#/components/schemas/Vendor"
...
    CreateMenuItemBody:
      type: object
      properties:
        vendorId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        isAvailable:
          type: boolean
      required: [vendorId, name, price, category]
    UpdateMenuItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        isAvailable:
          type: boolean
...