# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CampusEats API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  studentId: number;
  vendorId: number;
  studentName?: string;
  status: OrderStatus;
  totalAmount: number;
  pickupCode?: string;
  notes?: string;
  createdAt?: Date;
  updatedAt?: Date;
}